Create main.c
Main draw loop with the doom fire and a cube. 
Can't use the entire screen for the doom fire due to code size limitations. Cube calculations can be done better (precalc or sinetable) instead of realtime calculations.
Uses VSync flags to avoid tearing and the two layers defined are used as a double buffer mechanism.